+++ CHANGELOG.md
@@ -6,4 +6,22 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- Run current lua file on any DCS server according to address.
+- Run current lua file in either mission or GUI scripting environment.
+- Get mission theatre command for quick testing.
+- Settings to change server address, port, web auth credentials, https.
+- Display return in output window.
+
+## [1.0.0] - 2024-01-09
+
+### Added
+- Run currently selected code via right click menu.
+- Commands and buttons to quickly switch between local and remote DCS server
+- Commands and buttons to quickly switch between mission and GUI environment.
+- Warning for remote DCS server address not set.
+
+### Changed
+- Consolidate run code commands to different targets into one, execute based on current setting. 
+- Improve output to show current runner target.